[Neuron] Values should be 'double' and not 'int'

+++ core/engine.hpp
@@ -172,7 +172,7 @@ class engine {
 
   static vector<double> get_values(string variable) {
     vector<double> values(neuron_count());
-    for(vector<int>::size_type index = 0; index < values.size(); ++index) {
+    for(vector<double>::size_type index = 0; index < values.size(); ++index) {
       values[index] = neuron_value(index, variable);
     }
     return values;
